Improved login flow to prompt associations with existing accounts (#118)
* Improved login flow to prompt associations with existing accounts

If a new account's identity provider is associated with an email address
that is associated with other identity providers, the user is now prompted
to associate their new login with the existing identity-provider. This only
happens the first time; if they decline, then any subsequent logins will
skip that step, and are kept as a separate account.

Fixed a bug with getting backend errors out of query params. The code
was reading SearchParamState.error, which is a separate property that
usually returns undefined. Now getting the .value of the state before
reading the keyID.

Changed magiclink login flow to use ID as the /login identifier and token
is now passed as a query param, instead of the token being the identifier.
This should make it nearly impossible to brute-force guess a magiclink.

Resolves IR-3991

* Fixed formatting

* fixup

* fixup

---------

Co-authored-by: hanzlamateen <hanzlamateen@live.com>